pyproject.toml
[tool.poetry]
name = "checkbox"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "checkbox"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
black = "^24.2.0"
fastapi-filter = {extras = ["sqlalchemy"], version = "^1.1.0"}
fastapi-pagination = "^0.12.19"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
pytest = "^8.2.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
